+++ engines/python/setup/djl_python/properties_manager/hf_properties.py
@@ -137,9 +137,10 @@ class HuggingFaceProperties(Properties):
 
         # TODO remove this after refactor of all handlers
         # device map is not required for lmi dist and vllm
-        if properties['rolling_batch'] in {RollingBatchEnum.lmidist,
-                                           RollingBatchEnum.vllm,
-                                           RollingBatchEnum.lmidist_v2}:
+        if properties['rolling_batch'] in {
+                RollingBatchEnum.lmidist, RollingBatchEnum.vllm,
+                RollingBatchEnum.lmidist_v2
+        }:
             return properties
 
         if properties[
