+++ engines/python/setup/djl_python/properties_manager/vllm_rb_properties.py
@@ -59,7 +59,6 @@ class VllmRbProperties(Properties):
 
         return properties
 
-
     # TODO: Remove this once SageMaker resolved driver issue
     @root_validator(skip_on_failure=True)
     def set_eager_model_for_quantize(cls, properties):
